+++ example/android/app/src/main/java/com/dynamsoft/flutter_qrcode_scanner_example/MainActivity.java
@@ -1,4 +1,4 @@
-package com.dynamsoft.flutter_qrcode_scanner_example;
+package com.dynamsoft.flutter_camera_qrcode_scanner_example;
 
 import io.flutter.embedding.android.FlutterActivity;