+++ example/android/app/src/main/kotlin/com/dynamsoft/flutter_qrcode_scanner_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.dynamsoft.flutter_camera_qrcode_scanner_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}