+++ lib/flutter_camera_qrcode_scanner.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
@@ -26,15 +27,25 @@ class _ScannerViewState extends State<ScannerView> {
 
   @override
   Widget build(BuildContext context) {
-    const String viewType = 'com.dynamsoft.flutter_qrcode_scanner/nativeview';
+    const String viewType =
+        'com.dynamsoft.flutter_camera_qrcode_scanner/nativeview';
     final Map<String, dynamic> creationParams = <String, dynamic>{};
 
-    return AndroidView(
-      viewType: viewType,
-      onPlatformViewCreated: _onPlatformViewCreated,
-      creationParams: creationParams,
-      creationParamsCodec: const StandardMessageCodec(),
-    );
+    if (Platform.isAndroid) {
+      return AndroidView(
+        viewType: viewType,
+        onPlatformViewCreated: _onPlatformViewCreated,
+        creationParams: creationParams,
+        creationParamsCodec: const StandardMessageCodec(),
+      );
+    } else {
+      return UiKitView(
+        viewType: viewType,
+        onPlatformViewCreated: _onPlatformViewCreated,
+        creationParams: creationParams,
+        creationParamsCodec: const StandardMessageCodec(),
+      );
+    }
   }
 
   void _onPlatformViewCreated(int id) {
@@ -51,8 +62,8 @@ class ScannerViewController {
   Stream<List<BarcodeResult>> get scannedDataStream => _streamController.stream;
 
   ScannerViewController(int id) {
-    _channel =
-        MethodChannel('com.dynamsoft.flutter_qrcode_scanner/nativeview_$id');
+    _channel = MethodChannel(
+        'com.dynamsoft.flutter_camera_qrcode_scanner/nativeview_$id');
     _channel.setMethodCallHandler((call) async {
       switch (call.method) {
         case 'onDetected':