+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0
+* Support both Android and iOS development.
+* Update Dynamsoft Camera Enhancer to 2.1.1.
+* Update Dynamsoft Barcode Reader to 8.9.1.
+
 ## 0.0.1
 
 * A lightweight Flutter widget for QR code scanning.